# openapi.yaml
openapi: 3.0.1
info:
  title: Forward API
  description: |-
    REST and SOAP backend for ForwardService, the Ford x FIAP 2026 challenge platform. Exposes customer, vehicle, lead, churn score and service event resources for the dealer network.

    **Authentication**: every endpoint except `/health` and `/ready` requires a bearer JWT (Supabase) **or** an `X-API-Key` header for server-to-server callers (n8n, cron jobs).

    **Rate limiting**: requests are limited per IP and subject via Bucket4j. When the bucket is empty the API returns `429 Too Many Requests`.

    **Errors**: every error response follows RFC 7807 `application/problem+json` using Spring's `ProblemDetail` model. User-facing `detail` messages are in Portuguese (pt-BR).
  contact:
    name: Forward Team
    url: https://github.com/fwd-ford/forward-api-java
    email: dev@fwd-ford.com
  license:
    name: Proprietary
    url: https://github.com/fwd-ford/forward-api-java/blob/main/LICENSE
  version: 0.1.0
...
servers:
- url: http://localhost:18080
  description: Local Docker
- url: http://localhost:8080
  description: Local native (mvnw)
- url: https://forward-api.fwd-ford.com
  description: Production
security:
- bearerAuth: []
- apiKey: []
...
paths:
...
  /api/v1/customers/{id}:
    get:
      tags:
      - Customers
      summary: Get customer by id
      description: Returns the customer profile for the given UUID. Authentication
        is required; in Sprint 1 any authenticated caller may read any customer (the
        mobile app needs this for the Lead Detail flow — see fwd-ford/forward-api-java#23).
        Sprint 2 will tighten access to dealer-scoped reads via the leads table.
      operationId: getCustomer
      parameters:
      - name: id
        in: path
        description: Customer UUID.
        required: true
        schema:
          type: string
          format: uuid
        example: 2ddd2b47-9a80-4a0c-8c0a-8ee35d6f8b10
      responses:
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: Customer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "403":
          description: Forbidden by RBAC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Invalid UUID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
components:
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
...
    Customer:
      type: object
      properties:
        id:
          type: string
          description: Customer UUID.
          format: uuid
          example: 2ddd2b47-9a80-4a0c-8c0a-8ee35d6f8b10
        full_name:
          type: string
          description: Customer full name.
          example: Maria Aparecida Silva
        email:
          type: string
          description: Customer email (RFC 5322).
          format: email
          example: maria.silva@example.com
        phone:
          type: string
          description: Phone number with country and area code.
          example: "+5511987654321"
        city:
          type: string
          description: City of residence.
          example: Sao Paulo
        state:
          maxLength: 2
          minLength: 2
          type: string
          description: Brazilian state abbreviation (UF).
          example: SP
        opt_in_whatsapp:
          type: boolean
          description: True when the customer accepted WhatsApp outreach.
          example: true
        created_at:
          type: string
          description: Timestamp the customer record was created.
          format: date-time
          example: 2025-09-12T18:34:00-03:00
      description: "Customer profile, returned with RBAC enforcement."
  securitySchemes:
    bearerAuth:
      type: http
      description: Supabase JWT issued to authenticated users. Validated against the
        configured HS256 secret or the project JWKS (asymmetric).
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      description: "Internal API key for server-to-server calls (n8n, cron). Set via\
        \ the `INTERNAL_API_KEY` env var."
      name: X-API-Key
      in: header